refactor TransactionStatusNotification
