+++ packages/ui/src/common/components/TransactionStatusNotification/index.ts
@@ -1,0 +1,1 @@
+export * from './TransactionStatusNotification'
